[tests] Move dtests to src/integration package
In preparation for the work to improve M3 integration tests,
promote the existing integration tests (i.e. dtests) to a package
that makes more sense and is easier to discover.
