+++ src/integration/harness.go
@@ -1,5 +1,3 @@
-// +build dtest
-//
 // Copyright (c) 2020 Uber Technologies, Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -20,34 +18,5 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package harness
-
-import (
-	"fmt"
-	"os"
-	"testing"
-
-	"github.com/m3db/m3/src/cmd/tools/dtest/docker/harness/resources"
-)
-
-var singleDBNodeDockerResources resources.DockerResources
-
-func TestMain(m *testing.M) {
-	var err error
-	singleDBNodeDockerResources, err = resources.SetupSingleM3DBNode(
-		resources.WithExistingCluster("dbnode01", "coord01"),
-	)
-
-	if err != nil {
-		fmt.Println("could not set up db docker containers", err)
-		os.Exit(1)
-	}
-
-	if l := len(singleDBNodeDockerResources.Nodes()); l != 1 {
-		fmt.Println("should only have a single node, have", l)
-		os.Exit(1)
-	}
-
-	code := m.Run()
-	os.Exit(code)
-}
+// Package integration contains tests that are run against docker containers
+package integration
