+++ src/integration/harness_test.go
@@ -1,3 +1,5 @@
+// +build dtest
+//
 // Copyright (c) 2020 Uber Technologies, Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -18,5 +20,34 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-// Package harness contains tests that are run against docker containers
-package harness
+package integration
+
+import (
+	"fmt"
+	"os"
+	"testing"
+
+	"github.com/m3db/m3/src/integration/resources"
+)
+
+var singleDBNodeDockerResources resources.DockerResources
+
+func TestMain(m *testing.M) {
+	var err error
+	singleDBNodeDockerResources, err = resources.SetupSingleM3DBNode(
+		resources.WithExistingCluster("dbnode01", "coord01"),
+	)
+
+	if err != nil {
+		fmt.Println("could not set up db docker containers", err)
+		os.Exit(1)
+	}
+
+	if l := len(singleDBNodeDockerResources.Nodes()); l != 1 {
+		fmt.Println("should only have a single node, have", l)
+		os.Exit(1)
+	}
+
+	code := m.Run()
+	os.Exit(code)
+}
