+++ src/integration/query_api_test.go
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package harness
+package integration
 
 import (
 	"encoding/json"
@@ -28,9 +28,9 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
+	"github.com/m3db/m3/src/integration/resources"
 
-	"github.com/m3db/m3/src/cmd/tools/dtest/docker/harness/resources"
+	"github.com/stretchr/testify/assert"
 )
 
 type urlTest struct {
