+++ src/integration/resources/docker_resource.go
